Integration test suite.
This extracts the e2e tests as a separate module so that they can
be run separately, after the image has been built.
