+++ .github/workflows/pr-validation.yaml
@@ -127,6 +127,23 @@ jobs:
       - name: Build and load (current arch)
         run: |
           docker buildx build --load -t ${{ env.NAME }}:${{ steps.prep.outputs.VERSION }} .
+      - name: Setup go
+        uses: actions/setup-go@v2
+        with:
+          go-version: "1.16"
+      - name: Install libgit2
+        run: |
+          cd /tmp
+          git clone https://github.com/libgit2/libgit2.git
+          cd libgit2
+          git checkout maint/v1.1
+          mkdir build && cd build
+          cmake ..
+          sudo cmake --build . --target install
+          sudo ldconfig
+      - name: Run integration tests
+        run: |
+          make verify GITOPS_PROMOTION_IMAGE=${{ env.NAME }}:${{ steps.prep.outputs.VERSION }}
       - name: Run Trivy vulnerability scanner
         uses: aquasecurity/trivy-action@master
         with:
